Published Update 1.2.5
This adds several fixes, patches and bug fixes for pre-existing issues such as the custom road tilesheets for WttDR not making the stone sound when walking over them, the horse path not functioning at all or correctly even when re-enabled. NPC Pathing issues with Ridgeside Village and Stardew Valley Expanded will be fixed in Version 1.3.0.