Pushing changes for 1.3.4
Re-adding support for Fostoria and Tego's Stardew Valley Expanded, Updating all MapStrings in each tmx map file and all i18n translation files to start with WTD in order to prevent MapStrings conflicting with other mods such as Cape Stardew and the Cave map, also removed all mention of a "HorsePath" replaced it with mentions of "RoadSigns", I've moved various json files to the data folder in assets including all intermod compats for easier organisation and updates in the future, I removed the retro recolour entries as the tilesheets weren't changed in the slightest and I can't update the curvy roads in the condition they are in right now, I would need to pay someone to do that, whoever would be willing to. updating all tilesheet tile properties with the ones from vanilla maps, added vines/leaves to the Zuzu City highway sign for each season except Winter. Added two tiles to each season variant of the roads that has a bird on the right bottom post, thanks again to XxHarvzBackxX the Downtown Zuzu Mod Author for the Highway Signs and decorations. Updated manist.json update keys order, not sure if it'll do anything or not and removed the large ForestRoadSquare tmx map file and instead mad it a small tmx patch file to replace the bottom section of the road tiles since that's all that was different between the two maps anyways. Report any bugs you encounter to me on my Discord if you so please, however I'll probably find them first.