Pushing changes for 1.3.4
Added intermod compatibility for Lizzie's Stoffton Town mod, this will work by adding a new T-Section to the Winding Roads map where you can walk to the right and reach the town if you don't have the Fostoria mod installed, otherwise you won't see it. Uncommented the Eemie content because the recolour still technically exists, but I'm not going to do anything about it for now, I fixed some of the warps for the Stardew Valley Expanded mod compat with the Desert map and the Desert Festival, also updated the content.json format to 2.6.0.